Remove condition around 'touch' commands
I did this initially because the pre-existing code to create the
~/.zshrc file was wrapped in a condition checking for the existance of
the file. However I don't see why this condition is necessary. `touch`
will only update the accessed and modified timestamps on a file if it
already exists, and in all cases we are modifying the files at this
time.